Separate bin and lib
