+++ bin/flowtrace-show
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('../src/render').main();
+require('../src/bin/render').main();
