+++ src/index.js
@@ -1,10 +1,6 @@
 module.exports = {
   // eslint-disable-next-line global-require
-  trace: require('./trace'),
+  trace: require('./lib/trace'),
   // eslint-disable-next-line global-require
-  render: require('./render'),
-  // eslint-disable-next-line global-require
-  replay: require('./replay'),
-  // eslint-disable-next-line global-require
-  common: require('./common'),
+  Flowtrace: require('./lib/Flowtrace'),
 };
