+++ src/lib/common.js
