+++ src/lib/trace.js
